server/requirements.txt
# === Core ===
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# === Database ===
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
geoalchemy2>=0.14.0

# === Auth ===
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# === HTTP / External API ===
httpx>=0.27.0

# === Push Notification ===
firebase-admin>=6.4.0

# === Scheduler ===
apscheduler>=3.10.0

# === Cache ===
redis>=5.0.0

# === Storage ===
boto3>=1.34.0

# === Utilities ===
python-multipart>=0.0.9
pillow>=10.2.0
